refactor(hooks): move the skill router to its own opt-in PR
The UserPromptSubmit skill router is a separate behavioral feature from
profile carriers: it injects suggestions on every matching prompt and
needs its own precision/recall and latency evidence. It now lives on
feat/skill-router as an opt-in adapter; this branch carries only the
carrier generator.

Kept here: the run-with-flags.js fix that stops UserPromptSubmit hooks
from echoing raw stdin into model context when disabled, dry-run, or
missing, with a focused test (tests/hooks/run-with-flags-user-prompt.test.js).
