--- scripts/lib/skill-router.js
@@ -1,315 +1,0 @@
-/**
- * Skill routing for the UserPromptSubmit skill-router hook.
- *
- * Scores the user's prompt against skill frontmatter descriptions with plain
- * offline token matching and returns the best matches, marking each one as
- * installed in the active plugin or available on demand from the source
- * catalog. Slim profile plugins generated by scripts/plugin-profiles.js
- * record their source repository in ecc-profile.json at the plugin root, so
- * routing covers the FULL catalog even when only a slim profile is enabled.
- */
-
-'use strict';
-
-const crypto = require('crypto');
-const fs = require('fs');
-const os = require('os');
-const path = require('path');
-const { parseFrontmatter } = require('./plugin-profiles');
-
-const PROFILE_METADATA_FILE = 'ecc-profile.json';
-const CACHE_TTL_MS = 6 * 60 * 60 * 1000;
-const DEFAULT_MAX_RESULTS = 3;
-const DEFAULT_MIN_SCORE = 3;
-
-// Function words and generic task verbs only. Domain-ish words (test, fix,
-// build, review, ...) stay routable because skills are named after them.
-const STOPWORDS = new Set([
-  'the', 'and', 'for', 'with', 'from', 'this', 'that', 'these', 'those',
-  'have', 'has', 'had', 'you', 'your', 'our', 'are', 'was', 'were', 'been',
-  'being', 'will', 'would', 'could', 'should', 'can', 'may', 'might', 'must',
-  'not', 'but', 'into', 'onto', 'about', 'over', 'under', 'out', 'off',
-  'when', 'where', 'what', 'which', 'who', 'whom', 'why', 'how', 'all',
-  'any', 'some', 'each', 'more', 'most', 'other', 'such', 'only', 'also',
-  'than', 'then', 'them', 'they', 'there', 'here', 'just', 'like', 'very',
-  'really', 'please', 'help', 'want', 'need', 'make', 'made', 'using',
-  'use', 'used', 'get', 'got', 'set', 'let', 'way', 'thing', 'things',
-  'know', 'still', 'now', 'currently', 'something',
-]);
-
-/**
- * Lowercase, split on non-alphanumerics, drop stopwords and short tokens.
- * Long plurals also contribute their singular form so "hooks" matches "hook".
- */
-function tokenize(text) {
-  const tokens = new Set();
-  for (const rawToken of String(text || '').toLowerCase().split(/[^a-z0-9]+/)) {
-    if (rawToken.length < 3 || STOPWORDS.has(rawToken)) {
-      continue;
-    }
-    tokens.add(rawToken);
-    if (rawToken.length > 3 && rawToken.endsWith('s')) {
-      tokens.add(rawToken.slice(0, -1));
-    }
-  }
-  return tokens;
-}
-
-/**
- * List immediate skill directory names under a `skills/` root.
- *
- * @param {string} skillsRoot Absolute path to a `skills/` directory.
- * @returns {Array<string>} Sorted directory names, empty when absent.
- */
-function listSkillDirs(skillsRoot) {
-  if (!fs.existsSync(skillsRoot)) {
-    return [];
-  }
-  return fs.readdirSync(skillsRoot, { withFileTypes: true })
-    .filter(entry => entry.isDirectory())
-    .map(entry => entry.name)
-    .sort();
-}
-
-/**
- * Scan every SKILL.md under a source root for its id and description.
- *
- * @param {string} sourceRoot Root of an ECC checkout.
- * @returns {Array<{id: string, description: string}>} Catalog entries.
- */
-function readCatalog(sourceRoot) {
-  const skillsRoot = path.join(sourceRoot, 'skills');
-  const entries = [];
-  for (const skillId of listSkillDirs(skillsRoot)) {
-    const skillPath = path.join(skillsRoot, skillId, 'SKILL.md');
-    if (!fs.existsSync(skillPath)) {
-      continue;
-    }
-    const { description } = parseFrontmatter(fs.readFileSync(skillPath, 'utf8'));
-    entries.push({ id: skillId, description });
-  }
-  return entries;
-}
-
-/**
- * Cheap fingerprint used to decide whether a cached catalog is still valid.
- *
- * Tracks only skill directory count and the `skills/` mtime, so adds and
- * removes invalidate the cache while in-place description edits ride the
- * TTL — the tradeoff that keeps this inside the blocking-hook budget.
- *
- * @param {string} sourceRoot Root of an ECC checkout.
- * @returns {{dirCount: number, mtimeMs: number}} Signature for cache comparison.
- */
-function catalogSignature(sourceRoot) {
-  const skillsRoot = path.join(sourceRoot, 'skills');
-  try {
-    const stat = fs.statSync(skillsRoot);
-    return { dirCount: listSkillDirs(skillsRoot).length, mtimeMs: stat.mtimeMs };
-  } catch {
-    return { dirCount: 0, mtimeMs: 0 };
-  }
-}
-
-/**
- * Resolve the cache file backing one source root.
- *
- * The cache lives under the user's home directory, NOT os.tmpdir(): a
- * stable, predictable filename in a shared /tmp would let another local
- * user pre-plant cache content that gets injected into this user's context.
- *
- * @param {string} sourceRoot Root of an ECC checkout.
- * @returns {string} Absolute path to this root's cache file.
- */
-function cachePathFor(sourceRoot) {
-  const cacheDir = process.env.ECC_SKILL_ROUTER_CACHE_DIR
-    || path.join(os.homedir(), '.claude', 'cache');
-  const digest = crypto.createHash('sha1').update(path.resolve(sourceRoot)).digest('hex').slice(0, 12);
-  return path.join(cacheDir, `ecc-skill-router-${digest}.json`);
-}
-
-/**
- * Keep only well-formed {id, description} pairs.
- *
- * Applied to every catalog source that is not a fresh in-process scan — the
- * on-disk cache and a plugin's embedded snapshot are both attacker-writable
- * in the threat model, and a malformed entry would otherwise reach scoring
- * (where a non-string id throws) or routed output.
- *
- * @param {unknown} entries Candidate catalog array.
- * @returns {Array<{id: string, description: string}>} Validated entries.
- */
-function sanitizeCatalogEntries(entries) {
-  if (!Array.isArray(entries)) {
-    return [];
-  }
-  return entries
-    .filter(entry => entry
-      && typeof entry.id === 'string'
-      && entry.id.length > 0
-      && typeof entry.description === 'string')
-    .map(entry => ({ id: entry.id, description: entry.description }));
-}
-
-/**
- * Catalog scan with a best-effort cache under the user's home directory.
- * The signature only tracks skill directory adds/removes, so edited
- * descriptions can be stale for up to the TTL — acceptable for routing
- * hints, and it keeps the hook well under the blocking-hook latency budget.
- *
- * @param {string} sourceRoot Root of the ECC checkout holding `skills/`.
- * @returns {Array<{id: string, description: string}>} Catalog entries.
- */
-function loadCatalog(sourceRoot) {
-  const signature = catalogSignature(sourceRoot);
-  const cachePath = cachePathFor(sourceRoot);
-
-  try {
-    const cached = JSON.parse(fs.readFileSync(cachePath, 'utf8'));
-    if (cached.signature
-      && cached.signature.dirCount === signature.dirCount
-      && cached.signature.mtimeMs === signature.mtimeMs
-      && Number.isFinite(cached.builtAt)
-      && cached.builtAt <= Date.now()
-      && Date.now() - cached.builtAt < CACHE_TTL_MS
-      && Array.isArray(cached.entries)) {
-      return sanitizeCatalogEntries(cached.entries);
-    }
-  } catch {
-    // missing or unreadable cache: rebuild below
-  }
-
-  const entries = readCatalog(sourceRoot);
-  writeCatalogCache(cachePath, { signature, builtAt: Date.now(), entries });
-  return entries;
-}
-
-/**
- * Persist the catalog cache without ever writing through a planted symlink.
- *
- * The payload is written to an exclusive per-process temp file and renamed
- * over the target: `wx` fails rather than following an existing link, and
- * the rename is atomic, so there is no lstat-then-write window for another
- * local user to swap the path in between (TOCTOU). Best-effort throughout —
- * a cache that cannot be written just means the next run rescans.
- *
- * @param {string} cachePath Destination cache file.
- * @param {object} payload Cache document to serialize.
- * @returns {void}
- */
-function writeCatalogCache(cachePath, payload) {
-  const tempPath = `${cachePath}.${process.pid}.tmp`;
-  try {
-    fs.mkdirSync(path.dirname(cachePath), { recursive: true });
-    fs.writeFileSync(tempPath, JSON.stringify(payload), { mode: 0o600, flag: 'wx' });
-    fs.renameSync(tempPath, cachePath);
-  } catch {
-    try {
-      fs.unlinkSync(tempPath);
-    } catch {
-      // nothing to clean up
-    }
-  }
-}
-
-/**
- * Verify a candidate path fingerprints as a real ECC checkout.
- *
- * A sourceRoot from ecc-profile.json is only honored when it looks like a
- * real ECC checkout. The metadata is plugin-supplied data, so an arbitrary
- * path must never end up in routed output (Prompt Defense Baseline).
- *
- * @param {string} candidate Path supplied by plugin metadata.
- * @returns {boolean} Whether the path may be trusted as a catalog source.
- */
-function isEccSourceRoot(candidate) {
-  return fs.existsSync(path.join(candidate, 'skills'))
-    && fs.existsSync(path.join(candidate, 'manifests', 'install-modules.json'));
-}
-
-/**
- * Resolve where the full skill catalog lives and which skills are installed
- * in the active plugin. Generated slim profiles carry an ecc-profile.json
- * with their source repository and an embedded catalog snapshot; the full
- * plugin routes over itself. A missing or unverifiable sourceRoot falls
- * back to installed-only routing.
- */
-function resolveRouterContext(pluginRoot) {
-  const installedIds = new Set(listSkillDirs(path.join(pluginRoot, 'skills')));
-  let sourceRoot = pluginRoot;
-  let embeddedCatalog = null;
-
-  const metadataPath = path.join(pluginRoot, PROFILE_METADATA_FILE);
-  try {
-    const metadata = JSON.parse(fs.readFileSync(metadataPath, 'utf8'));
-    if (metadata && typeof metadata.sourceRoot === 'string' && isEccSourceRoot(metadata.sourceRoot)) {
-      sourceRoot = metadata.sourceRoot;
-      if (Array.isArray(metadata.catalog)) {
-        embeddedCatalog = sanitizeCatalogEntries(metadata.catalog);
-      }
-    }
-  } catch {
-    // no profile metadata: plugin root is the catalog source
-  }
-
-  return { sourceRoot, installedIds, embeddedCatalog };
-}
-
-/**
- * Score the prompt against the catalog. Skill-id token matches weigh 3,
- * description token matches weigh 1; ties break alphabetically so output
- * is deterministic.
- */
-function routePrompt(prompt, options = {}) {
-  const pluginRoot = options.pluginRoot;
-  if (!pluginRoot) {
-    throw new Error('routePrompt requires pluginRoot');
-  }
-  const maxResults = options.maxResults ?? DEFAULT_MAX_RESULTS;
-  const minScore = options.minScore ?? DEFAULT_MIN_SCORE;
-
-  const promptTokens = tokenize(prompt);
-  if (promptTokens.size === 0) {
-    return [];
-  }
-
-  const { sourceRoot, installedIds, embeddedCatalog } = resolveRouterContext(pluginRoot);
-  const scored = [];
-
-  for (const entry of embeddedCatalog || loadCatalog(sourceRoot)) {
-    if (entry.id === 'ecc-catalog') {
-      continue;
-    }
-    let score = 0;
-    for (const token of tokenize(entry.id.replace(/-/g, ' '))) {
-      if (promptTokens.has(token)) {
-        score += 3;
-      }
-    }
-    for (const token of tokenize(entry.description)) {
-      if (promptTokens.has(token)) {
-        score += 1;
-      }
-    }
-    if (score >= minScore) {
-      scored.push({
-        id: entry.id,
-        description: entry.description,
-        score,
-        installed: installedIds.has(entry.id),
-        sourceRoot,
-      });
-    }
-  }
-
-  scored.sort((a, b) => b.score - a.score || a.id.localeCompare(b.id));
-  return scored.slice(0, maxResults);
-}
-
-module.exports = {
-  PROFILE_METADATA_FILE,
-  tokenize,
-  sanitizeCatalogEntries,
-  resolveRouterContext,
-  routePrompt,
-};
